refactored notifications
