+++ src/main/java/UseCases/Notification/PriceDropNotification.java
@@ -1,29 +1,26 @@
-package Controller;
+package UseCases.Notification;
+import Controller.Scheduler;
 import Entities.*;
-import java.io.IOException;
 import java.util.TimerTask;
 
+/** A price drop notification use case that tells us if a product drops below the user's desired price */
 public class PriceDropNotification implements BaseNotification {
     private final Scheduler scheduler;
-    private final Item item;
+    private final Product product;
     private Boolean showNotification;
 
-    public PriceDropNotification(Item item) {
+    public PriceDropNotification(Product product) {
         this.showNotification = Boolean.FALSE;
         TimerTask checkSale = new TimerTask() {
             @Override
             public void run() {
-                try {
-                    item.updatePrice();
-                    checkNotification();
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
-                }
+                // TODO: Add refactored item search api call
+                checkNotification();
             }
         };
 
         this.scheduler = new Scheduler(checkSale, 1000 * 60 * 60 * 24);
-        this.item = item;
+        this.product = product;
     }
     public Boolean getShowNotification() {
         return showNotification;
@@ -41,9 +38,10 @@ public class PriceDropNotification implements BaseNotification {
         this.scheduler.disableTimer();
     }
 
-    /** Logic to check if price drop notification should be seen */
+    /** Logic to check if price drop notification should be seen
+     * @returns boolean if notification should be shown */
     public boolean checkNotification() {
-        this.showNotification = this.item.isPriceBelowDesiredPrice();
+        this.showNotification = this.product.getProductPrice() <= this.product.getProductDesiredPrice();
         return this.showNotification;
     }
 }
