+++ src/main/java/UseCases/Notification/SaleNotification.java
@@ -1,29 +1,26 @@
-package Controller;
+package UseCases.Notification;
+import Controller.Scheduler;
 import Entities.*;
-import java.io.IOException;
 import java.util.TimerTask;
 
+/** A sale notification use case that tells us if a product is on sale */
 public class SaleNotification implements BaseNotification {
     private final Scheduler scheduler;
-    private final Item item;
+    private final Product product;
     private Boolean showNotification;
 
-    public SaleNotification(Item item) {
+    public SaleNotification(Product product) {
         this.showNotification = Boolean.FALSE;
         TimerTask checkSale = new TimerTask() {
             @Override
             public void run() {
-                try {
-                    item.updatePrice();
-                    checkNotification();
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
-                }
+                // TODO: Add refactored item search api call
+                checkNotification();
             }
         };
 
         this.scheduler = new Scheduler(checkSale, 1000 * 60 * 60 * 24);
-        this.item = item;
+        this.product = product;
     }
 
     public Boolean getShowNotification() {
@@ -42,9 +39,10 @@ public class SaleNotification implements BaseNotification {
         this.scheduler.disableTimer();
     }
 
-    /** Logic to check if sale notification should be seen */
+    /** Logic to check if sale notification should be seen
+     * @returns boolean if notification should be shown */
     public boolean checkNotification() {
-        this.showNotification = item.isItemOnSale();
+        this.showNotification = product.getPriceChange() < 0;
         return this.showNotification;
     }
 }
